change restCliente verHabitacion()
Se ha cambiado el tipo de parámetro que hay que pasarle al método verHabitacion() de float a int.
Y ahora sí saca la información de la API.

+++ API/ProyectoDAMROOMS/src/main/java/ies/mz/ProyectoDAMROOMS/microservicios/RestClient.java
@@ -32,7 +32,7 @@ public class RestClient {
         }
     }
 
-    public void verHabitacion(float numero){
+    public void verHabitacion(int numero){
         String resultado = this.client.target("http://localhost:8080/habitaciones/"+numero)
                 .request(MediaType.APPLICATION_JSON)
                 .accept(MediaType.APPLICATION_JSON)
